+++ src/avr/read_only_memory.rs
@@ -1,3 +1,3 @@
 pub trait ReadOnlyMemory {
-  fn read(&self, address: u16) -> u16;
+  fn read(&self, address: u32) -> u16;
 }
